// src/api/generated/models/infoStatusCreateRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Community App V2
 * OpenAPI spec version: 0.1.0
 */
import type { InfoStatus } from './infoStatus'

/**
 * Append one simple status message and update the current status.
 */
export interface InfoStatusCreateRequest {
  status: InfoStatus
  message?: string | null
}